App casi finalizada

+++ src/main/java/com/project/controllers/UsuarioController.java
@@ -29,7 +29,6 @@ public class UsuarioController {
 
 	@Autowired
 	UsuarioRepository repository;
-		
 
 	UsuarioController(UsuarioRepository repository) {
 		this.repository = repository;
@@ -45,10 +44,24 @@ public class UsuarioController {
 	}
 	// end::get-aggregate-root[]
 
+	@CrossOrigin
 	@PostMapping("/usuarios/new")
 	@ResponseBody
 	UserHector newUser(@RequestBody UserHector newUser) {
-		return repository.save(newUser);
+		List<UserHector> users = repository.findAll();
+		boolean encontrado = false;
+		for (UserHector userHector : users) {
+			if (userHector.getEmail().equals(newUser.getEmail())) {
+				encontrado = true;
+			}
+		}
+
+		if (!encontrado) {
+			return repository.save(newUser);
+		} else {
+			throw new ResponseStatusException(HttpStatus.UNPROCESSABLE_ENTITY, "El usuario ya existe");
+		}
+
 	}
 
 	// Single item
@@ -60,12 +73,15 @@ public class UsuarioController {
 		return repository.findById(id).orElseThrow(() -> new UserNotFoundException(id));
 	}
 
-	@PutMapping("/usuarios/{id}")
+	@CrossOrigin
+	@PutMapping("/usuarios/update/{id}")
 	@ResponseBody
 	UserHector replaceUser(@RequestBody UserHector newUser, @PathVariable int id) {
 
 		return repository.findById(id).map(user -> {
 			user.setEmail(newUser.getEmail());
+			user.setNombre(newUser.getNombre());
+			user.setApellidos(newUser.getApellidos());
 			user.setPassword(newUser.getPassword());
 			return repository.save(user);
 		}).orElseGet(() -> {
@@ -74,12 +90,24 @@ public class UsuarioController {
 		});
 	}
 
+	@CrossOrigin
 	@DeleteMapping("/usuarios/delete/{id}")
 	@ResponseBody
-	void deleteUser(@PathVariable int id) {
-		repository.deleteById(id);
+	ResponseEntity<UserHector> deleteUser(@PathVariable int id) {
+		final String USER_EMAIL = "admin@admin.es";
+		Optional<UserHector> userFromDatabase = repository.findById(id);
+		if (userFromDatabase.isPresent()) {
+			if (userFromDatabase.get().getEmail().equals(USER_EMAIL)) {
+				return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+			}
+			repository.deleteById(id);
+			return new ResponseEntity<>(HttpStatus.OK);
+		}
+		return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
 	}
 
+	@CrossOrigin
 	@PostMapping("/usuarios/login")
 	public ResponseEntity<UserHector> login(@RequestBody UserHector user) {
 
@@ -87,20 +115,19 @@ public class UsuarioController {
 				.orElseThrow(() -> new ResponseStatusException(HttpStatus.UNPROCESSABLE_ENTITY, "Login incorrecto"));
 		if (!userInDatabase.getPassword().equals(user.getPassword())) {
 			throw new ResponseStatusException(HttpStatus.UNPROCESSABLE_ENTITY, "Login incorrecto");
-			
+
 		}
 		return new ResponseEntity<UserHector>(userInDatabase, HttpStatus.OK);
 	}
-	
-	//@PostMapping("/usuarios/registro")
-	
+
+	// @PostMapping("/usuarios/registro")
 
 	@PostMapping("/usuarios/signup")
 	public ResponseEntity<UserHector> signUp(@RequestBody UserHector user) {
 		return new ResponseEntity<UserHector>(repository.save(user), HttpStatus.OK);
 	}
 
-	public Optional<UserHector> buscarPorEmal( String email) {
+	public Optional<UserHector> buscarPorEmal(String email) {
 		return repository.findByEmail(email);
 	}
